pyproject.toml
# 
#   Muna
#   Copyright © 2025 NatML Inc. All Rights Reserved.
#

[project]
name = "muna"
dynamic = ["version"]
description = "Run AI models in Python. Register at https://muna.ai."
readme = "README.md"
dependencies = [
    "numpy",
    "pillow",
    "pydantic>=2.0",
    "requests",
    "rich",
    "typer"
]
requires-python = ">=3.10"
authors = [ { name = "NatML Inc.", email = "hi@muna.ai" } ]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Software Development :: Libraries",
]

[project.urls]
Homepage = "https://muna.ai"
Documentation = "https://docs.muna.ai"
Source = "https://github.com/muna-ai/muna-py"
Changelog = "https://github.com/muna-ai/muna-py/blob/main/Changelog.md"

[project.optional-dependencies]
# None

[project.scripts]
muna = "muna.cli.__init__:app"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["muna", "muna*"]
namespaces = false

[tool.setuptools.package-data]
"muna.lib" = ["macos/*/*.dylib", "windows/*/*.dll", "linux/*/*.so"]

[tool.setuptools.dynamic]
version = { attr = "muna.version.__version__" }